Halo regions moved to their own file.
